feat: v1.8.3-3
1. netkeeper-db.sh added update_password().
2. netkeeper-db.sh will ask for this choice when calling delete_all().

+++ files/netkeeper-db.sh
@@ -5,6 +5,7 @@ NETKEEPER_DB_PATH='/usr/lib/netkeeper/netkeeper.db'
 help() {
     echo "usage: $0 [command]"
     echo "       next_data"
+    echo "       update_password"
     echo "       select_all"
     echo "       delete_all"
     exit 0
@@ -18,6 +19,14 @@ select * from netkeeper order by id limit 1;
 EOF
 }
 
+update_password() {
+    echo -n "Please enter your new password: "
+    read -r NEW_PASSWORD
+    if [ -n "$NEW_PASSWORD" ]; then
+        sqlite3 "$NETKEEPER_DB_PATH" "update netkeeper set password = $NEW_PASSWORD;"
+    fi
+}
+
 select_all() {
     sqlite3 "$NETKEEPER_DB_PATH" << EOF
 .header on
@@ -27,7 +36,12 @@ EOF
 }
 
 delete_all() {
-    sqlite3 "$NETKEEPER_DB_PATH" "delete from netkeeper;"
+    echo -n "Are you sure you want to delete all the content in the table? [y/N]"
+    read -r DEL_CHOOSE
+    case "$DEL_CHOOSE" in
+        [yY] | [yY][eE][sS])
+            sqlite3 "$NETKEEPER_DB_PATH" "delete from netkeeper;";;
+    esac
 }
 
 main() {
